feat: generate component documentation for mintlify
* feat: implement component documentation generator compatible with mintlify
* feat: improve mintlify integration
* chore: add a way to run codegen scripts standalone
* chore: remove vitepress dependency
