+++ alfred/run.py
@@ -6,7 +6,7 @@ import alfred
 @alfred.command("run.docs", help="preview the documentation as a developper")
 def run_docs():
     os.chdir('docs')
-    alfred.run("npm run dev")
+    alfred.run("npm run preview")
 
 @alfred.command("run.storybook", help="preview the storybook as a developper")
 def run_storybook():
