--- docs/docs/.vitepress/theme/custom.css
@@ -1,14 +1,0 @@
-:root {
-  --vp-c-brand: #29cf00;
-  --vp-c-brand-light: #91e74e;
-}
-
-main img {
-  border-radius: 8px;
-  min-width: 100%;
-}
-
-.VPHome .vp-doc.container {
-  padding: 0;
-  max-width: 100%;
-}
