--- docs/docs/.vitepress/theme/index.js
@@ -1,5 +1,0 @@
-import DefaultTheme from "vitepress/theme";
-import "./custom.css";
-import "highlight.js/styles/a11y-light.min.css";
-
-export default DefaultTheme;
