+++ docs/docs/components/.gitignore
@@ -1,3 +1,2 @@
 *
-!component-list.md
-!component_page.md.tpl
+!component_page.mdx.tpl
